Tests fail solution

+++ database/migrations/2024_05_25_120624_create_media_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('url');
             $table->enum('type', ['image', 'video', 'audio']);
             $table->string('alt')->nullable();
-            $table->integer('mediable_id');
+            $table->unsignedBigInteger('mediable_id');
             $table->string('mediable_type');
             $table->timestamps();
         });
